+++ src/main/java/io/bluebeaker/mtepatches/mixin/ic2/MixinTileEntityCrop.java
@@ -8,10 +8,14 @@ import net.minecraft.util.math.BlockPos;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.Random;
 
 @Mixin(value = TileEntityCrop.class,remap = false)
 public abstract class MixinTileEntityCrop {
+
     @Inject(method = "onNeighborChange",at = @At(value = "INVOKE", shift = At.Shift.AFTER, target = "Lic2/core/block/TileEntityBlock;onNeighborChange(Lnet/minecraft/block/Block;Lnet/minecraft/util/math/BlockPos;)V"),cancellable = true)
     public void onNeighborChange(Block neighbor, BlockPos neighborPos, CallbackInfo ci){
         if(!MTEPatchesConfig.ic2.cropOnAllFarmlands)
@@ -19,5 +23,11 @@ public abstract class MixinTileEntityCrop {
         if (((TileEntityCrop)(Object)this).getWorld().getBlockState(((TileEntityCrop)(Object)this).getPos().down()).getBlock() instanceof BlockFarmland){
             ci.cancel();
         }
+    }
+
+    @Redirect(method = "onEntityCollision",at = @At(value = "INVOKE", target = "Ljava/util/Random;nextInt(I)I"))
+    public int preventTrample(Random instance, int i){
+        if(MTEPatchesConfig.ic2.noTrampleCrops) return 1;
+        return instance.nextInt(i);
     }
 }